authenticate and authorizitaion added for jwt

+++ controllers/api/v1/post_api.js
@@ -12,23 +12,22 @@ module.exports.index = async function (req, res) {
   });
 };
 
-
 module.exports.destroyPost = async function (req, res) {
   try {
     const post = await Post.findById(req.query.id);
-    // if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.query.id });
+      await Comment.deleteMany({ post: req.query.id });
 
-    return res.json(200, {
-      message: "Post Deleted Successfully",
-    });
-    // } else {
-    //   req.flash("error", "You can't Deleted a post");
-
-    //   return res.redirect("back");
-    // }
+      return res.json(200, {
+        message: "Post Deleted Successfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "Yoy can't delete this post",
+      });
+    }
   } catch (error) {
     console.log("********", error);
     return res.json(500, {
